requirements.txt
jwcrypto
six
python-etcd
